+++ Changelog.md
@@ -1,5 +1,6 @@
 # SupraFit 2.0 pre-Alpha
 
+- add variable peak time for integration in thermogram dialog
 - export integrated peaks from thermogram dialog directly to txt file
 - add splitter widgets into thermogram dialog, make chart and table resizable, stores last geometry
 - add calibration peak for ITC